Add MCP shared module and Cloudflare worker transport
