+++ package.json
@@ -32,6 +32,7 @@
     "bun": ">=1.2.0"
   },
   "dependencies": {
+    "@cfworker/json-schema": "^4.1.0",
     "@modelcontextprotocol/sdk": "^1.25.1",
     "three": "^0.177.0",
     "zod": "^4.2.1"
